+++ src/tests/streams/transform/BUILD.bazel
@@ -1,0 +1,18 @@
+load("//:build/wd_test.bzl", "wd_test")
+
+# The transform streams suite runs the same test modules under two configs:
+# transform-cpp against the C++ implementation, transform-ts against the
+# TypeScript implementation.
+
+transform_suite_srcs = glob(["*.js"]) + ["transform-modules.capnp"]
+
+wd_test(
+    src = "transform-cpp.wd-test",
+    data = transform_suite_srcs,
+)
+
+wd_test(
+    src = "transform-ts.wd-test",
+    args = ["--experimental"],
+    data = transform_suite_srcs,
+)
